requirements.txt
azure-functions
pandas
azure-storage-file-datalake
